Crate fxaclient_ffi

Structs

ACCOUNTS

Functions

fxa_authorize_auth_code⚠
Try to get an OAuth code using a session token.

fxa_begin_oauth_flow⚠
Request a OAuth token by starting a new OAuth flow.

fxa_begin_pairing_flow⚠
Request a OAuth token by starting a new pairing flow, by calling the content server pairing endpoint.

fxa_bytebuffer_free

fxa_check_authorization_status
Retrieve the refresh token authorization status.

fxa_clear_access_token_cache
This method should be called when a request made with an OAuth token failed with an authentication error. It clears the internal cache of OAuth access tokens, so the caller can try to call fxa_get_access_token or fxa_profile again.

fxa_complete_oauth_flow
Finish an OAuth flow initiated by [fxa_begin_oauth_flow].

fxa_disconnect
Disconnect from the account and optionaly destroy our device record.

fxa_ensure_capabilities⚠
Ensure that the device capabilities are registered with the server.

fxa_free

fxa_from_json
Restore a [FirefoxAccount] instance from an serialized state (created with [fxa_to_json]).

fxa_gather_telemetry
Gather telemetry collected by FxA.

fxa_get_access_token
Try to get an access token.

fxa_get_connection_success_url
Get the url to redirect after there has been a successful connection to FxA.

fxa_get_current_device_id
Try to get the current device id from state.

fxa_get_devices
Fetch the devices (including the current one) in the current account.

fxa_get_ecosystem_anon_id

fxa_get_manage_account_url
Get the url to open the user's account-management page.

fxa_get_manage_devices_url
Get the url to open the user's devices-management page.

fxa_get_pairing_authority_url
Get the pairing URL to navigate to on the Auth side (typically a computer).

fxa_get_session_token
Try to get a session token.

fxa_get_token_server_endpoint_url
Get the Sync token server endpoint URL.

fxa_handle_push_message
Handle a push payload coming from the Firefox Account servers.

fxa_handle_session_token_change
Typically called during a password change flow. Invalidate all tokens and get a new refresh token.

fxa_initialize_device⚠
Initalizes our own device, most of the time this will be called right after logging-in for the first time.

fxa_is_in_migration_state
Check if there is migration state.

fxa_migrate_from_session_token
Migrate from a logged-in sessionToken Firefox Account.

fxa_new
Creates a [FirefoxAccount].

fxa_poll_device_commands
Poll and parse available remote commands targeted to our own device.

fxa_profile
Fetches the profile associated with a Firefox Account.

fxa_retry_migrate_from_session_token
Retry the migration attempt using the stored migration state.

fxa_send_tab
Send a tab to another device identified by its Device ID.

fxa_set_device_name
Update the display name for the current device.

fxa_set_push_subscription
Update the Push subscription information for the current device.

fxa_str_free⚠
Public destructor for strings managed by the other side of the FFI.

fxa_to_json
Serializes the state of a [FirefoxAccount] instance. It can be restored later with [fxa_from_json].